refactor(quality): implement P1/P3/P5a quality improvements from issue #862
P1 — Break core↔status circular dependency:
- Extract `core/identity_aliases.py` into new leaf package
  `specify_cli.identity.aliases` (no deps on core or status)
- Update `status/models.py` and `status/progress.py` to import from
  `specify_cli.identity` instead of `specify_cli.core`
- Replace `core/identity_aliases.py` with a backward-compat re-export shim
- Verified: `from specify_cli.status.models import Lane` no longer
  transitively imports from `specify_cli.core`

P3 — Remove false deprecation from tasks_support:
- Move all logic to `specify_cli.task_utils.support` (new package)
- Replace `tasks_support.py` with a thin re-export shim
- Update all 27 src/ callers to import from `specify_cli.task_utils`
- Fix ruff issues in the moved code (unused imports, B904 chain)

P5a — Move auth global state out of __init__.py:
- Extract `_tm`, `_tm_lock`, `get_token_manager()`, `reset_token_manager()`
  from `auth/__init__.py` into `auth/manager.py`
- `auth/__init__.py` is now a thin re-export with no mutable globals
- Update 3 test fixtures to patch `specify_cli.auth.manager._tm` instead
  of the now-defunct `specify_cli.auth._tm`

All existing tests pass. Architectural layer tests pass.
