+++ src/specify_cli/auth/__init__.py
@@ -14,15 +14,12 @@ Public API:
 - The auth error hierarchy — re-exported so downstream code imports errors
   from ``specify_cli.auth`` instead of flow-internal modules.
 
-This module deliberately uses a module-level ``_tm`` + ``threading.Lock``
-rather than a class attribute or singleton pattern — decision from the WP01
-review of the previous run, where a class-level singleton caused lifecycle
-ambiguity and made testing difficult.
+The singleton state lives in :mod:`specify_cli.auth.manager` to avoid
+mutable globals at package-init time. Double-checked locking rationale is
+documented there.
 """
 
 from __future__ import annotations
-
-import threading
 
 from .errors import (
     AuthenticationError,
@@ -44,6 +41,7 @@ from .errors import (
     StorageDecryptionError,
     TokenRefreshError,
 )
+from .manager import get_token_manager, reset_token_manager
 from .secure_storage import SecureStorage
 from .token_manager import TokenManager
 
@@ -74,34 +72,3 @@ __all__ = [
     "StorageBackendUnavailableError",
     "StorageDecryptionError",
 ]
-
-_tm: TokenManager | None = None
-_tm_lock = threading.Lock()
-
-
-def get_token_manager() -> TokenManager:
-    """Return the process-wide :class:`TokenManager` instance.
-
-    Lazy-initializes from secure storage on first call; subsequent calls
-    return the same instance. Thread-safe via double-checked locking.
-    """
-    global _tm
-    if _tm is None:
-        with _tm_lock:
-            if _tm is None:
-                storage = SecureStorage.from_environment()
-                tm = TokenManager(storage)
-                tm.load_from_storage_sync()
-                _tm = tm
-    return _tm
-
-
-def reset_token_manager() -> None:
-    """Reset the global :class:`TokenManager`. For tests only.
-
-    Used by ``tests/auth/conftest.py`` to guarantee state isolation between
-    test cases.
-    """
-    global _tm
-    with _tm_lock:
-        _tm = None
