+++ src/specify_cli/cli/commands/invocations_cmd.py
@@ -36,7 +36,7 @@ import typer
 from rich.console import Console
 from rich.table import Table
 
-from specify_cli.tasks_support import find_repo_root
+from specify_cli.task_utils import find_repo_root
 
 app = typer.Typer(name="invocations", help="Query local invocation records.")
 console = Console()
