+++ src/specify_cli/cli/commands/profiles_cmd.py
@@ -9,7 +9,7 @@ from rich.console import Console
 from rich.table import Table
 
 from specify_cli.invocation.registry import ProfileRegistry
-from specify_cli.tasks_support import find_repo_root
+from specify_cli.task_utils import find_repo_root
 
 app = typer.Typer(name="profiles", help="Manage and list agent profiles.")
 console = Console()
