+++ src/specify_cli/identity/__init__.py
@@ -1,4 +1,4 @@
-"""Backward-compat shim — canonical home is specify_cli.identity.aliases."""
+"""Identity compatibility helpers — leaf package with no dependencies on core or status."""
 
 from specify_cli.identity.aliases import with_tracked_mission_slug_aliases
