+++ src/specify_cli/status/progress.py
@@ -16,7 +16,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
 
-from specify_cli.core.identity_aliases import with_tracked_mission_slug_aliases
+from specify_cli.identity.aliases import with_tracked_mission_slug_aliases
 from specify_cli.mission_metadata import mission_identity_fields
 
 from .models import Lane, StatusSnapshot
