+++ src/specify_cli/task_utils/__init__.py
@@ -1,0 +1,56 @@
+"""Work package utilities for spec-kitty task manipulation.
+
+Re-exports the public surface of task_utils.support for convenience.
+"""
+
+from specify_cli.task_utils.support import (
+    LANE_ALIASES,
+    LANES,
+    TIMESTAMP_FORMAT,
+    TaskCliError,
+    WorkPackage,
+    activity_entries,
+    append_activity_log,
+    build_document,
+    detect_conflicting_wp_status,
+    ensure_lane,
+    extract_scalar,
+    find_repo_root,
+    get_lane_from_frontmatter,
+    git_status_lines,
+    is_legacy_format,
+    load_meta,
+    locate_work_package,
+    match_frontmatter_line,
+    normalize_note,
+    now_utc,
+    run_git,
+    set_scalar,
+    split_frontmatter,
+)
+
+__all__ = [
+    "LANE_ALIASES",
+    "LANES",
+    "TIMESTAMP_FORMAT",
+    "TaskCliError",
+    "WorkPackage",
+    "activity_entries",
+    "append_activity_log",
+    "build_document",
+    "detect_conflicting_wp_status",
+    "ensure_lane",
+    "extract_scalar",
+    "find_repo_root",
+    "get_lane_from_frontmatter",
+    "git_status_lines",
+    "is_legacy_format",
+    "load_meta",
+    "locate_work_package",
+    "match_frontmatter_line",
+    "normalize_note",
+    "now_utc",
+    "run_git",
+    "set_scalar",
+    "split_frontmatter",
+]
